python/pyproject.toml
[project]
name = "tierkreis"
description = "Python client and utilities for tierkreis."
readme = "README.md"
dynamic = ["version"]
authors = [
    { name = "Seyon Sivarajah", email = "seyon.sivarajah@quantinuum.com" },
    { name = "Lukas Heidemann", email = "lukas.heidemann@quantinuum.com" },
    { name = "John Children", email = "john.children@quantinuum.com" },
    { name = "Alan Lawrence", email = "alan.lawrence@quantinuum.com" },
]

requires-python = ">=3.10,<3.13"
dependencies = [
    "betterproto[compiler]==2.0.0b6",
    "grpclib>=0.4.3rc,<0.5",          # pre-release to support python 3.10
    "networkx>=2.6.3,<3",
    "graphviz>=0.20,<0.21",
    "click==8.1.3",                   # 8.1.4 causes mypy fails
    "yachalk>=0.1.4,<0.2",
    "requests>=2.31,<3",
    "pydantic~=2.5",
]

[project.scripts]
tkrs = 'tierkreis.cli:cli'

[project.optional-dependencies]
docker = ["docker>=5,<6"]

telemetry = [
    "opentelemetry-sdk>=1.15.0,<2",
    "opentelemetry-exporter-otlp>=1.15.0,<2",
]

commontypes = ["pytket>=1.0"]
lint = [
    "ruff~=0.3",
    "pyright==1.1.345",
    "tierkreis[sc22-example,telemetry,docker]",
]
test = [
    "pytest>=6.2,<7",
    "pytest-asyncio>=0.16,<0.17",
    "pytest-cov>=5.0,<6",
    "tierkreis[commontypes,typecheck]",
]
sc22-example = ["numpy>=1.20,<2", "pytket>=1.0"]

typecheck = ["tierkreis_typecheck>=0.2,<0.3"]


docs = ["sphinx>=4.3", "sphinx-book-theme>=1.1.2"]

[tool.pytest.ini_options]
markers = ["pytket"]


[build-system]
requires = ["setuptools>=65.5.0", "betterproto[compiler]==2.0.0b6"]
build-backend = "generate_protos"
backend-path = ["tierkreis/_build"]

[tool.setuptools.dynamic]
version = { attr = "tierkreis._version.__version__" }

[tool.setuptools.packages.find]
include = ["tierkreis*", "tierkreis/py.typed"]


[tool.ruff]
target-version = "py310"
# default + imports
lint.select = ["E4", "E7", "E9", "F", "I"]
extend-exclude = ["tierkreis/core/protos"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # module imported but unused

[tool.pyright]
include = ["."]
exclude = ["tierkreis/core/protos", "build"]
ignore = ["^build/", "pytket_worker"]
